+++ generators/python/src/fern_python/generators/sdk/client_generator/client_generator.py
@@ -120,14 +120,15 @@ class ClientGenerator:
                     self._is_default_body_parameter_used = True
 
                 if generated_endpoint_function.snippet is not None:
-                    if is_async:
-                        self._snippet_registry.register_async_client_endpoint_snippet(
-                            endpoint=endpoint, expr=generated_endpoint_function.snippet
-                        )
-                    else:
-                        self._snippet_registry.register_sync_client_endpoint_snippet(
-                            endpoint=endpoint, expr=generated_endpoint_function.snippet
-                        )
+                    for snippet in generated_endpoint_function.snippet:
+                        if is_async:
+                            self._snippet_registry.register_async_client_endpoint_snippet(
+                                endpoint=endpoint, expr=snippet.snippet, example_id=snippet.example_id
+                            )
+                        else:
+                            self._snippet_registry.register_sync_client_endpoint_snippet(
+                                endpoint=endpoint, expr=snippet.snippet, example_id=snippet.example_id
+                            )
 
         return class_declaration