+++ generators/python/src/fern_python/snippet/snippet_endpoint_expression.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from typing import Optional
 
 import fern.generator_exec.resources as generator_exec
 
@@ -7,5 +8,6 @@ from fern_python.codegen import AST
 
 @dataclass
 class EndpointExpression:
+    example_id: Optional[str]
     endpoint_id: generator_exec.EndpointIdentifier
     expr: AST.Expression