+++ generators/go/sdk/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## Unreleased -->
 
+## [0.22.1 - 2024-05-23]
+
+- Internal: SDK snippets now specify their example ID and we create a snippet per example
+
 ## [0.22.0 - 2024-05-21]
 
 - Feature: Extra properties decoded from response objects are retained and accessible via the