+++ generators/typescript/sdk/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.20.1-rc0] - 2024-05-23
+
+- Internal: SDK snippets now specify their example ID and we create a snippet per example
+
 ## [0.20.0-rc1] - 2024-05-20
 
 - Fix: Pass `abortSignal` to `Stream` for server-sent-events and JSON streams so that the user